rename test -> tests and docs -> doc
