+++ doc/source/references.rst
