+++ doc/source/veerer_demo.rst
