+++ tests/test_angles.py
@@ -13,14 +13,11 @@
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
 #
-# You should have received a copy of the GNU General Public License
+# Yoa should have received a copy of the GNU General Public License
 # along with veerer. If not, see <https://www.gnu.org/licenses/>.
 ######################################################################
 
-import sys
 import pytest
-
-import random
 from veerer import VeeringTriangulation, RED, BLUE, PURPLE, GREEN
 
 @pytest.mark.parametrize("fp, cols",
@@ -34,7 +31,21 @@ from veerer import VeeringTriangulation, RED, BLUE, PURPLE, GREEN
 ])
 def test_triangle(fp, cols):
     V = VeeringTriangulation(fp, cols)
-    A = V.abelian_cover()
-    assert A.angles() == [2], (V, A)
+    assert V.angles() == [1, 1, 1, 1], V
 
-if __name__ == '__main__': sys.exit(pytest.main(sys.argv))
+@pytest.mark.parametrize("fp, cols",
+[
+ ("(0,1,2)(~0,~1,~2)", "RRB"),
+ ("(0,1,2)(~0,~1,~2)", "BBR"),
+ ("(0,1,2)(~0,~1,~2)", "PBR"),
+ ("(0,1,2)(~0,~1,~2)", "GRB"),
+ ("(0,1,2)(~0,~1,~2)", "RPG"),
+ ("(0,1,2)(~0,~1,~2)", "BGP")
+])
+def test_torus(fp, cols):
+    V = VeeringTriangulation(fp, cols)
+    assert V.angles() == [2], V
+
+if __name__ == '__main__':
+    import sys
+    sys.exit(pytest.main(sys.argv))
