+++ tests/test_flip_sequence.py
@@ -17,7 +17,6 @@
 # along with veerer. If not, see <https://www.gnu.org/licenses/>.
 ######################################################################
 
-import sys
 import pytest
 
 import random
@@ -82,4 +81,6 @@ def test_Q_11_p5():
         h = g.inverse()
         assert f.is_identical(h), (f, h)
 
-if __name__ == '__main__': sys.exit(pytest.main(sys.argv))
+if __name__ == '__main__':
+    import sys
+    sys.exit(pytest.main(sys.argv))
