+++ tests/test_triangulation_comparison.py
@@ -17,7 +17,6 @@
 # along with veerer. If not, see <https://www.gnu.org/licenses/>.
 ######################################################################
 
-import sys
 import pytest
 
 from veerer import Triangulation, VeeringTriangulation
@@ -52,4 +51,6 @@ def test_veering_triangulation_comparison():
     assert V1 != V4
     assert V3 != V4
 
-if __name__ == '__main__': sys.exit(pytest.main(sys.argv))
+if __name__ == '__main__':
+    import sys
+    sys.exit(pytest.main(sys.argv))
